refactor: Cards 일급컬렉션 추가
- Participant 인터페이스 제거
- Cards 추가에 따른 테스트 코드 변경
