--- src/main/java/domain/Participant.java
@@ -1,9 +1,0 @@
-package domain;
-
-
-public interface Participant {
-
-    boolean addOneCard(Card card);
-
-    int sumCardNumbers();
-}
